Add timeshift_dump binary utility
When recording dumps, it can take a long time from start of the recording
to actually record packets. This initial wait is potentially unwanted when
replaying the recording, so we want to make a new dump file with times
shifted by the wait.

Refactor Inner::write_packet function to a pub fn, so it can be used
in a binary. Add elapsed parameter.

Also:
- Change dump formatting a bit (spacing), to conform to the dump files
  already recorded by another project.